requirements.txt
Boruta>=0.3.0
scikit-learn>=1.3.0,<1.6.0
pandas>=2.0.0,<2.2.0
numpy>=1.26.0,<2.0.0
shap>=0.42.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tqdm>=4.66.0
